Added Windows service hosting

+++ src/DotNetHole/Program.cs
@@ -5,6 +5,7 @@ using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
+using System.ServiceProcess;
 using System.Threading;
 using System.Threading.Tasks;
 using ARSoft.Tools.Net.Dns;
@@ -13,51 +14,101 @@ namespace DotNetHole
 {
 	public class Program
 	{
-		public static async Task Main()
+		public static void Main(string[] args)
 		{
-			await new ProgramInstance().Run();
-			Process.GetCurrentProcess().Kill();
+			var dotNetHoleService = new DotNetHoleService();
+
+			if (!IsConsoleWindow())
+			{
+				ServiceBase.Run(dotNetHoleService);
+				return;
+			}
+
+			dotNetHoleService.Start();
+
+			Console.CancelKeyPress += (sender, eventArgs) =>
+			{
+				dotNetHoleService.Stop();
+				Process.GetCurrentProcess().Kill();
+			};
+
+			while (true)
+				Console.ReadKey();
+		}
+
+		private static bool IsConsoleWindow()
+		{
+			try
+			{
+				var _ = Console.WindowWidth;
+				return true;
+			}
+			catch (IOException)
+			{
+				return false;
+			}
 		}
 	}
 
-	public class ProgramInstance
+	public class DotNetHoleService : ServiceBase
 	{
 		private const bool SimpleLogging = true;
 		private DnsClient _client;
 
-		private readonly TaskCompletionSource<object> _taskCompletionSource = new TaskCompletionSource<object>();
 		private readonly HashSet<string> _blacklistedUrls = new HashSet<string>();
 		private readonly SemaphoreSlim _logSemaphore = new SemaphoreSlim(1, 1);
 
 		private long _blockCount;
 		private long _totalCount;
+		private DnsServer _server;
+		private bool _isService;
 
-		public async Task Run()
+		protected override async void OnStart(string[] args)
+		{
+			_isService = true;
+			await Start();
+			base.OnStart(args);
+		}
+
+		protected override void OnStop()
+		{
+			_server.Stop();
+			base.OnStop();
+		}
+
+		protected override void OnPause()
+		{
+			_server.Stop();
+			base.OnPause();
+		}
+
+		protected override void OnContinue()
+		{
+			_server.Start();
+			base.OnContinue();
+		}
+
+		public async Task Start()
 		{
 			await LoadBlacklist();
 
-			Console.WriteLine($"Blacklisted domains: {_blacklistedUrls.Count}");
+			LogToConsole($"Blacklisted domains: {_blacklistedUrls.Count}");
 
 			_client = new DnsClient(IPAddress.Parse("1.1.1.1"), 200);
 
-			Console.Clear();
-
-			var server = new DnsServer(10, 10);
-			server.QueryReceived += ServerOnQueryReceived;
-			server.Start();
-
-			Console.CancelKeyPress += (sender, eventArgs) =>
-			{
-				server.Stop();
-				_taskCompletionSource.SetResult(null);
-			};
-
-			await _taskCompletionSource.Task;
+			_server = new DnsServer(10, 10);
+			_server.QueryReceived += ServerOnQueryReceived;
+			_server.Start();
 		}
 
 		private async Task LoadBlacklist()
 		{
-			var blacklistFileUrls = await File.ReadAllLinesAsync("blacklists.txt");
+			var applicationDirectory = Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName);
+			var blacklistsFilePath = Path.Combine(applicationDirectory, "blacklists.txt");
+			if (!File.Exists(blacklistsFilePath))
+				throw new FileNotFoundException($"Blacklists file not found at path: {blacklistsFilePath}");
+
+			var blacklistFileUrls = await File.ReadAllLinesAsync(blacklistsFilePath);
 			using var httpClient = new HttpClient();
 
 			foreach (var url in blacklistFileUrls)
@@ -65,11 +116,11 @@ namespace DotNetHole
 				string blacklistContent;
 				try
 				{
-					blacklistContent = (await httpClient.GetStringAsync(url));
+					blacklistContent = await httpClient.GetStringAsync(url);
 				}
 				catch (Exception)
 				{
-					Console.WriteLine($"Blacklist {url} failed to load.");
+					LogToConsole($"Blacklist {url} failed to load.", true);
 					continue;
 				}
 
@@ -90,24 +141,22 @@ namespace DotNetHole
 
 		private async Task ServerOnQueryReceived(object sender, QueryReceivedEventArgs eventargs)
 		{
-			if (!(eventargs.Query is DnsMessage dnsMessage))
-			{
-				return;
-			}
+			if (!(eventargs.Query is DnsMessage dnsMessage)) return;
 
 			var isBlacklisted = IsBlacklisted(dnsMessage);
 			var response = await _client.SendMessageAsync(dnsMessage);
 
 			if (!isBlacklisted)
-			{
 				eventargs.Response = response;
-			}
 			else
 			{
 				var newMessage = ReplaceARecords(response, dnsMessage);
 
 				eventargs.Response = newMessage;
 			}
+
+			if (_isService)
+				return;
 
 #pragma warning disable 4014
 			LogQuery(dnsMessage, (DnsMessage) eventargs.Response, isBlacklisted);
@@ -166,46 +215,51 @@ namespace DotNetHole
 			if (isBlacklisted)
 				_blockCount++;
 
-
 			if (SimpleLogging)
 				LogQuerySimple();
 			else
 				LogQueryFull(questionMessage, responseMessage, isBlacklisted);
-
 
 			_logSemaphore.Release();
 		}
 
 		private void LogQuerySimple()
 		{
-			var percentage = Math.Round((double)_blockCount / _totalCount * 100, 2);
+			var percentage = Math.Round((double) _blockCount / _totalCount * 100, 2);
 			var logMessage = $"Block count: {_blockCount}/{_totalCount} ({percentage}%)";
 
-			Console.Title = logMessage;
-
-			Console.SetCursorPosition(0, 0);
-			Console.WriteLine($"{logMessage}          ");
+			LogToConsole($"{logMessage}          ", false, true);
 		}
 
-		private static void LogQueryFull(DnsMessage questionMessage, DnsMessage responseMessage, bool isBlacklisted)
+		private void LogQueryFull(DnsMessage questionMessage, DnsMessage responseMessage, bool isBlacklisted)
 		{
-			if (isBlacklisted || responseMessage == null)
-			{
-				Console.ForegroundColor = ConsoleColor.Red;
-			}
-
-			Console.WriteLine(questionMessage.Questions.First().Name);
+			LogToConsole(questionMessage.Questions.First().Name, red: isBlacklisted);
 			foreach (var answer in responseMessage.AnswerRecords)
 			{
 				switch (answer)
 				{
 					case AddressRecordBase addressRecordBase:
-						Console.WriteLine($"\t{addressRecordBase.Address}");
+						LogToConsole($"\t{addressRecordBase.Address}", red: isBlacklisted);
 						break;
 				}
 			}
+		}
 
-			Console.ResetColor();
+		private void LogToConsole(object message, bool red = false, bool overwrite = false)
+		{
+			if (_isService)
+				return;
+
+			if (red)
+				Console.ForegroundColor = ConsoleColor.Red;
+
+			if (overwrite)
+				Console.SetCursorPosition(0, 0);
+
+			Console.WriteLine(message);
+
+			if (red)
+				Console.ResetColor();
 		}
 	}
 }
